Lets create a controller set for  objects we are going to call items
most componets are build around the idea of table or set of tables of
data and managing that data. so to make this sample component useful
instead of just hello world style I am going to assume that your
componet wants to have a list of  objects and we are going to call them
items

so we create a pural and a singual controller, this way even though we
are creating more files it makes the code cleaner and easier to
understand. The singualar is dealing with one record, like and edit
form, and the pural is dealing with the list and filtering views.
These can be merged to one file but we are going not going too.