Creating the administrator models to pair up with our controllers
Again, we have two models one for sigual and one for pural usage. This
is a key difference because we are extending different base classes in
this case. the signular extends JModelAdmin, because this has features
built in for  unpublishing and publishing toggle buttons, getting the
form data from our Jforms xml files. The Pural has JModelList as a base
because it is designed for displaying admin lists.