adding in the Table file to tell our componet and joomla about our MYSQL table
Every single table you  create in your database needs to have a file to
tell the application about the table.

So in our case we are creating a file, that  tells our component about
our items  table.  We don't need to define  anything exept the name and
the primary key in this case it is very standard table with a primary
key called ID that is int, and auto incrementing. JTables are very
useful down the line and can do a lot of heavy lifting for you  esp. if
you have  advanced things you need to do to the table before  or on
save or   want to do validation before any  commits are stored to the
tables.