adding sql files
in your sql directory you will now fine the install and uninstall mysql
comments.

In this example I have DROP TABLE IF exists, and to uninstall the table
 on component uninstall some component develoepers  have a practice of
leaving in mysql tables after installs which can be helpful sometimes,
but it can also leave a bunch of unused tables I think when components
unistall so should the tables from the database if you want you can
remove those lines.