Ok we have mysql table, a controller, a model and a jtable we are going to now need a JFORM xml file
Now I have mixed feelings about JForm. Sometimes it makes simple jobs
over complicated and sometimes to make  hard jobs easier.

It isn't going anywhere, and you can override the forms with plugins so
I am going to stick with it here for completeness and for standards
sake.

 You can spend a lot of time reading up on JFORM because basically all
you are doing is  trying to generate forms   for your fields with XML
files.