Adding List View
Now to make the singual and list views, fromt he Controllers and the
models we already created we need to add in our views. the views are
what the users actually see  so take care to make them neat and nice so
that user has a good time interacting with your component.  Now that
Joomla team has spent much time on the JUI, you can create mobile ready
advanced components without too much  in the UI department.  Here is
the list view